assets/pycode/nexi-backend-master/requirements.txt
PyYAML==6.0
watchdog==2.1.9
